+++ packages/nrg/index.js
@@ -21,10 +21,11 @@ const {
 const {
   enrichAndLogError,
   addErrorToResponse,
-  handleError
+  handleError,
+  testError
 } = require('./lib/middleware/error')
 
-const { addToResponse, redirect } = require('./lib/middleware/end')
+const { addToResponse, redirect, noContent } = require('./lib/middleware/end')
 
 const { serveStatic, logClientMessage } = require('./lib/middleware/client')
 
@@ -60,7 +61,6 @@ const {
   getAccount,
   reduceAccountForClient,
   validateAccountUpdate,
-  validatePasswordUpdate,
   startEmailUpdate,
   updatePassword,
   updateAccount
@@ -119,6 +119,7 @@ module.exports = {
   enrichAndLogError,
   addErrorToResponse,
   handleError,
+  testError,
 
   // Serving:
   serveStatic,
@@ -126,6 +127,7 @@ module.exports = {
   // Result:
   addToResponse,
   redirect,
+  noContent,
 
   // Token:
   generateToken,
@@ -254,13 +256,11 @@ module.exports = {
   ],
 
   // Account Update:
-  validatePasswordUpdate,
   validateAccountUpdate,
   startEmailUpdate,
   updateAccount,
   accountUpdate: [
     requireAuthorization,
-    validatePasswordUpdate,
     validateAccountUpdate,
     getAccount,
     comparePasswords,