+++ packages/nrg/lib/middleware/error.js
@@ -3,6 +3,9 @@ function enrichAndLogError (err, ctx) {
   err.status = err.status || 500
   err.logLevel = err.logLevel || (err.status >= 500 ? 'error' : 'warn')
 
+  // Add request ID to error for external services like Sentry.
+  err.requestId = ctx.req.id
+
   // Log error.
   ctx.logger[err.logLevel](err)
 }
@@ -21,4 +24,24 @@ async function handleError (ctx, next) {
   }
 }
 
-module.exports = { enrichAndLogError, addErrorToResponse, handleError }
+class TestError extends Error {
+  constructor () {
+    super('This is just a test')
+    this.name = 'TestError'
+    this.logLevel = 'warn'
+  }
+}
+
+const secret = process.env.ERROR_TEST_SECRET
+function testError (ctx, next) {
+  if (secret && ctx.query.secret === secret) throw new TestError()
+  ctx.logger.warn('testError • Secret mismatch', { secret, query: ctx.query })
+  return next()
+}
+
+module.exports = {
+  enrichAndLogError,
+  addErrorToResponse,
+  handleError,
+  testError
+}