+++ packages/nrg-logger/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ianwalter/nrg-logger
 
+## 0.5.0
+
+### Minor Changes
+
+- b0a7c26: Adding debug for response body, add support for LogEntry severity, add support for logging ip addresses
+
+## 0.4.0
+
+### Minor Changes
+
+- 9e598b4: Adding req logger
+
+## 0.3.3
+
+### Patch Changes
+
+- 903a7a1: Update dependency @generates/logger to v1
+
 ## 0.3.2
 
 ### Patch Changes