+++ packages/nrg-mq/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ianwalter/nrg-mq
 
+## 1.2.8
+
+### Patch Changes
+
+- 09cc27b: Update dependency amqplib to ^0.8.0
+
+## 1.2.7
+
+### Patch Changes
+
+- 903a7a1: Update dependency @generates/logger to v1
+
 ## 1.2.6
 
 ### Patch Changes